+++ src/main/resources/schema.sql
@@ -1,13 +1,13 @@
 CREATE TABLE account (
-  id     INTEGER IDENTITY PRIMARY KEY,
+  id     BIGINT IDENTITY PRIMARY KEY,
   name   VARCHAR(255),
   amount DECIMAL(20, 2)
 );
 
 CREATE TABLE payment (
-  id                   INTEGER IDENTITY PRIMARY KEY,
-  SourceAccountId      INTEGER,
-  DestinationAccountId INTEGER,
+  id                   BIGINT IDENTITY PRIMARY KEY,
+  SourceAccountId      BIGINT,
+  DestinationAccountId BIGINT,
   amount               DECIMAL(20, 2),
   Description          VARCHAR(255),
   transactionDate      DATETIME
